# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Motus API",
        "description": "API REST pour le jeu de devinettes de mots Motus. Authentification par session cookie (Sanctum SPA).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Lister les parties de l'utilisateur courant",
                "operationId": "93736f40a135111a310e4b25ad61f30b",
                "responses": {
                    "200": {
                        "description": "Liste paginée des parties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedGames"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Démarrer une nouvelle partie",
                "operationId": "b8b9557618eb0f7b3650bdd4ab3a6487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "difficulty"
                                ],
                                "properties": {
                                    "difficulty": {
                                        "type": "string",
                                        "example": "easy",
                                        "enum": [
                                            "easy",
                                            "medium",
                                            "hard"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partie créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Difficulté invalide"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/games/{game}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Détail d'une partie",
                "operationId": "18467eed630a90558e542ccea443776a",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cette partie appartient à un autre utilisateur"
                    },
                    "404": {
                        "description": "Partie introuvable"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Attempt": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "attempt_number": {
                        "type": "integer",
                        "example": 1
                    },
                    "attempted_word": {
                        "type": "string",
                        "example": "MAISON"
                    },
                    "result": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "correct",
                                "present",
                                "absent"
                            ]
                        }
                    },
                    "is_correct": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Game": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "difficulty": {
                        "type": "string",
                        "example": "easy",
                        "nullable": true,
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ]
                    },
                    "length": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "first_letter": {
                        "type": "string",
                        "example": "M",
                        "nullable": true
                    },
                    "attempts_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "max_attempts": {
                        "type": "integer",
                        "example": 6
                    },
                    "status": {
                        "type": "string",
                        "example": "in_progress",
                        "enum": [
                            "in_progress",
                            "won",
                            "lost"
                        ]
                    },
                    "score": {
                        "type": "integer",
                        "example": 0
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "attempts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Attempt"
                        }
                    },
                    "word": {
                        "description": "Le mot cible. Null tant que la partie n'est pas terminée, révélé quand la partie est gagnée ou perdue.",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatedGames": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Game"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 20
                            },
                            "total": {
                                "type": "integer",
                                "example": 42
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}